Feat/binaries

+++ src/pipeline/extractor.rs
@@ -57,7 +57,7 @@ impl ExtractorPipe {
             let dir_name = entry.file_name().to_str().unwrap().to_string();
             let meta = metadata(entry.path())
                 .map_err(|e| ExecutionError::JobExecutionFailed(e.to_string(), e.to_string()))?;
-            if !mapped_str.contains(&dir_name) && meta.is_dir() {
+            if !mapped_str.contains(&dir_name) && meta.is_dir() && dir_name != ".bin" {
                 fs::remove_dir_all(entry.path()).await.map_err(|e| {
                     ExecutionError::JobExecutionFailed(e.to_string(), e.to_string())
                 })?;
